Add mixins for Layer2, Neighbor Discovery, and System functionalities; refactor routing mixin for improved organization
