+++ nornir_srl/connections/routing.py
@@ -1,13 +1,14 @@
-# BGP related methods extracted from srlinux.py
+# Routing related methods extracted from srlinux.py
 from __future__ import annotations
 
 from typing import Any, Dict, List, Optional
 import copy
 import jmespath
+from .helpers import lpm
 
 
-class BgpMixin:
-    """Mixin providing BGP related getters."""
+class RoutingMixin:
+    """Mixin providing routing and BGP related getters."""
 
     capabilities: Optional[Dict[str, Any]]
 
@@ -397,3 +398,162 @@ class BgpMixin:
         augment_resp(resp)
         res = jmespath.search(path_spec["jmespath"], resp[0])
         return {"bgp_peers": res}
+
+    def get_rib(
+        self,
+        afi: str,
+        network_instance: Optional[str] = "*",
+        lpm_address: Optional[str] = None,
+    ) -> Dict[str, Any]:
+        path_spec = {
+            "path": f"/network-instance[name={network_instance}]/route-table/{afi}",
+            "jmespath": '"network-instance"[?_hasrib].{NI:name, Rib:"route-table"."'
+            + afi
+            + '".route[].{"Prefix":"'
+            + ("ipv4-prefix" if afi == "ipv4-unicast" else "ipv6-prefix")
+            + '",\
+                    "next-hop":"_next-hop",type:"route-type", Act:active, "orig-vrf":"_orig_vrf",metric:metric, pref:preference, itf:"_nh_itf"}}',
+            "datatype": "state",
+        }
+
+        nhgroups = self.get(
+            paths=[
+                f"/network-instance[name={network_instance}]/route-table/next-hop-group[index=*]"
+            ],
+            datatype="state",
+        )
+        nhs = self.get(
+            paths=[
+                f"/network-instance[name={network_instance}]/route-table/next-hop[index=*]"
+            ],
+            datatype="state",
+        )
+
+        nh_mapping: Dict[str, Dict[str, Any]] = {}
+        for ni in nhs[0].get("network-instance", {}):
+            tmp_map: Dict[str, Any] = {}
+            for nh in ni["route-table"]["next-hop"]:
+                tmp_map[nh["index"]] = {
+                    "ip-address": nh.get("ip-address"),
+                    "type": nh.get("type"),
+                    "subinterface": nh.get("subinterface"),
+                }
+                if "resolving-tunnel" in nh:
+                    tmp_map[nh["index"]].update(
+                        {
+                            "tunnel": (nh.get("resolving-tunnel")).get("tunnel-type")
+                            + ":"
+                            + (nh.get("resolving-tunnel")).get("ip-prefix")
+                        }
+                    )
+                if "resolving-route" in nh:
+                    tmp_map[nh["index"]].update(
+                        {
+                            "resolving-route": (nh.get("resolving-route")).get(
+                                "ip-prefix"
+                            )
+                        }
+                    )
+
+            nh_mapping.update({ni["name"]: tmp_map})
+        nhgroup_mapping: Dict[str, Dict[str, List[Any]]] = {}
+        for ni in nhgroups[0].get("network-instance", {}):
+            network_instance = ni["name"]
+            nh_map: Dict[str, List[Any]] = {}
+            for nhgroup in ni["route-table"]["next-hop-group"]:
+                nh_map[nhgroup["index"]] = [
+                    nh_mapping[network_instance][nh.get("next-hop")]
+                    for nh in nhgroup.get("next-hop", [])
+                ]
+            nhgroup_mapping.update({ni["name"]: nh_map})
+
+        resp = self.get(
+            paths=[path_spec.get("path", "")], datatype=path_spec["datatype"]
+        )
+        for ni in resp[0].get("network-instance", {}):
+            if len(ni["route-table"][afi]) == 0:
+                ni["_hasrib"] = False
+            else:
+                ni["_hasrib"] = True
+                if lpm_address:
+                    lpm_prefix = lpm(
+                        lpm_address,
+                        [
+                            route[
+                                (
+                                    "ipv4-prefix"
+                                    if afi == "ipv4-unicast"
+                                    else "ipv6-prefix"
+                                )
+                            ]
+                            for route in ni["route-table"][afi]["route"]
+                        ],
+                    )
+                    if lpm_prefix:
+                        ni["route-table"][afi]["route"] = [
+                            r
+                            for r in ni["route-table"][afi]["route"]
+                            if r[
+                                (
+                                    "ipv4-prefix"
+                                    if afi == "ipv4-unicast"
+                                    else "ipv6-prefix"
+                                )
+                            ]
+                            == lpm_prefix
+                        ]
+                    else:
+                        ni["route-table"][afi]["route"] = []
+                        ni["_hasrib"] = False
+                        continue
+                for route in ni["route-table"][afi]["route"]:
+                    if route["active"]:
+                        route["active"] = "yes"
+                    else:
+                        route["active"] = "no"
+                    if "next-hop-group" in route:
+                        leaked = False
+                        if "origin-network-instance" in route:
+                            nh_ni = route["origin-network-instance"]
+                            if nh_ni != ni["name"]:
+                                leaked = True
+                                route["_orig_vrf"] = nh_ni
+                        else:
+                            nh_ni = ni["name"]
+                        route["_next-hop"] = [
+                            nh.get("ip-address")
+                            for nh in nhgroup_mapping[nh_ni].get(
+                                route["next-hop-group"], {}
+                            )
+                        ]
+
+                        route["_nh_itf"] = [
+                            (
+                                nh.get("subinterface") + f"@vrf:{nh_ni}"
+                                if leaked
+                                else nh.get("subinterface")
+                            )
+                            for nh in nhgroup_mapping[nh_ni].get(
+                                route["next-hop-group"], {}
+                            )
+                            if nh.get("subinterface")
+                        ]
+                        if len(route["_nh_itf"]) == 0:
+                            route["_nh_itf"] = [
+                                nh.get("tunnel")
+                                for nh in nhgroup_mapping[nh_ni].get(
+                                    route["next-hop-group"], {}
+                                )
+                                if nh.get("tunnel")
+                            ]
+                        if len(route["_nh_itf"]) == 0:
+                            resolving_routes = [
+                                nh.get("resolving-route", {})
+                                for nh in nhgroup_mapping[nh_ni].get(
+                                    route["next-hop-group"], {}
+                                )
+                                if nh.get("resolving-route")
+                            ]
+
+        res = jmespath.search(path_spec["jmespath"], resp[0])
+        return {"ip_rib": res}
